+++ mshell-crates/mshell-settings/src/lib.rs
@@ -52,6 +52,7 @@ mod monitors_settings;
 mod net;
 mod network_settings;
 mod notification_settings;
+mod osd_settings;
 mod overview_settings;
 mod plugins_settings;
 mod power_settings;